Launcher that supports proxying requests to the TVS server.
Extend Oak launcher to support proxying requests to TVS Server.
The part is written in C++ and statically linked to the rust code
Eventually the whole launcher will be rewritten in C++.

This launcher inherits all the functionality from Oak's launcher.

I'm patching Oak's qemu library to allow forwarding requests from the VMM to the proxying service.
These ports are hardcoded for now (similar to the kv-server ports), as there is no clean
way for me to stash the new service to the launcher service port.

Bug: b/344012112
Change-Id: Icf155e85a25cdd9b848d4b8c52e170d67f5e07f8