Orchestrator that talks to Hats launcher.
The delta with Oak's counterpart is that it talks to TVS via the launcher and fetches the token.
After talking to the Oak team it seems the best way forward for us is to use their orchestrator library.
However, there is a lot of logic in the orchestrator main, which I copied after taking their permission.

Note, the Grpc client seems complicated due to the fact that tonic grpc does not support bidirectional streaming
grpc very well. It assumes that the streams are asynchronus. In fact, if you don't use special synchronization
the client reads incoming messages first, and after the incoming stream is empty it starts sending messages.

I had to use multiple channels to synchronize.

Bug: b/344012112
Change-Id: Ia4a83a19a6b8b77f2c439df115efa3776c5733d1