Add certificates rust lib to fetch vcek
Launcher will provide a service calling the rust library in this PR to
fetch vcek. We use curl + rustls to fetch the VCEK from AMD website and
that's tested working well locally. The parameters depends on x86 CPUID
instruction, and /dev/sev being available. So this is very specific to
x86_64 and AMD SEV machine.
More work is expected for TDX.
Most functions are from snphost and sev ( apache 2.0 license ).

Change-Id: I22c138a8796ccbb119c351fe5f3a483b7b6ed5fe