Remove "external/<package_name> when the path has the package name.
Unfortunately, for some packages we need to use "external/package_name" prefix
in the include path to make it easier to know that the file came from a certain package.

Many packages make some efforts to have the path includes the package name, and for those
we remove the external prefix.

Bug: b/344012112
Change-Id: Id63363ca5ad065dda6cc5546f6968427a0508a70