Introduce a secondary handshake key to the TVS to facilitate key rotation.
The TVS takes an optional secondary key to be used in the handshake protocol.
In the session initiation, the client sends public key it intends to use along with
the handshake message.

The public key is used as a prologue to ensure both client & server have the same view of the state.

The idea of putting the public key in the prologue came after discussion with shabsi@.

Bug: b/344012112
Change-Id: Idc365368b5ca523b4b7de01042a5580062610045