Pass `secret` as an array of u8 bytes instead of string.
Apparently, rust string doesn't like certain non utf-8 strings.
Previously, we were passing the secret in the flag as hex of string.
Shortly, we will read secrets (full or partial HPKE private keys) from spanner
and we don't want to convert them to hex.

Here, we are having the test cli client to convert the returned secret to hex for debugging
and the rest of the code (orchestrator and TVS) deals with raw bytes.

Bug: b/356391320
Change-Id: Id9eeeeaee618f1112239b4ca34a325982a8dbf09