Create SecretData to hold keys, the object ensures proper wipeout of secrets and reduce leakage.
I looked around and the closest thing to what we want is in Tink library, however, the object is so complicated
and tries to solve another problem (it marks memory containing sensitive data so coredump redacts them) - I just
want to make sure that data are erased properly and disallow copying it by mistake.

SecretData is an object to store sensitive data. The object helps minimizing data leakage, and
ensure proper erasure that are often caused by optimization.
to other functions. The usage of this object reduce the need to explicitly
case/convert between data types as it can be constructed from strings and
string_view's, it also return a pointer to the inner array to allow manual
manipulation and to be passed to functions accepting arrays. Also, it returns
a string_view to be passed to other functions accepting strings. This is a
replacement to strings and vectors as the This is a replacement to strings
and vectors as the underlying data might be copied to other location. It is
not trivial to ensure proper erasure as implementation varies across
platform.

I will work on using this objects in all places that handles sensitive data.

Bug: b/344012112
Change-Id: I480c6016f4421666b7fa3f237f36798360982de0