aead-crypto: helper methods to generate keys, encrypt and decrypt messages using AES-256 in GCM mode.
We will use AEAD when developing an envelop encryption scheme to store secrets in Spanner in the next CL.

Encrypt and Decrypt takes a string for associated data, and we use two different strings.
one for TVS private keys, and another for secrets (full or partial HPKE private keys).

Encrypt returns nonce, ciphertext and tag in the same block in the following order: nonce|ciphertext|tag.
Decrypt expect the cipher text to have nonce, cihpertext and tag.
Further, the methods accept SecretData for keys and plain texts.

Bug: b/344012112
Change-Id: Id052acdb5bfff9e2b56be7762d215b9ef2adb434