Create a wrapper to key-fetcher and have TVS rust code calls the fetcher directly in GCP mode.
Instead of obtaining and decrypting the secret before calling the trusted code, we pass the key-fetcher
to the trusted code and if and only if the client passes verification, we fetch the secret, decrypt it and return it.

This is in alignment with what we would do if the trusted part of TVS runs in a low TCB enclave as it would ask the untrusted part(launcher)
to obtain information from Spanner.

I had to use exceptions to indicate errors as CXX only understands exception. I will work on creating a shim layer in CXX to convert C++ StatusOr
to rust Result.

Bug: b/344012112
Change-Id: I29eb8badaeabbdbb5a6a8c2842887e90d2c55434