Use posix_spawn in qemu.cc, add an option to allow outbound network connection, and few comsetic changes.
The main change in this CL is to use posix_spawn as system() doesn't like our command line arguments
since it relies on /bin/sh or the default shell to parse the argument (actually it launches a shell to launch qemu).

There are not to many great options to spawn binaries in C/C++:
* fork/execvp() is to low level, need some carefulness with descriptors, and not portable.
* popen shares the same problem as system().
* virtCommand() is the best solution; however, not sure if we want to link against their code for just one function.

I tested this CL and was able to launch qemu in default and SEV-SNP correctly (Es and Sev will be fixed soon).
I duplicated stdout and stderr for qemu to a temporary file instead of spamming the screen.
I also added an option to enable debugging over a TCP port.

Also, added an option to allow the CVM to have outbound network connection (next CL we will give the CVM a routable IP).

I did some cosmetic, readability in the CL that are not worth mentioning.

Bug: b/344012112
Change-Id: Ia9443997ba3bec56d3e3162755885554f4daf650