QEMU: add an option to give the CVM a routable IP address.
Use TAP and bridges to give the CVM a routable IP address.
In this mode, the CVM can be accessed from the host, and possibly outside.
The host can enable IP forwarding or it can add the bridge to the ethernet interface
so that the CVM gets IP address from the DHCP in the VLAN.

TAP requires some setup before starting a VM.
For now, we are asking the user to provide the bridge to use, IP address for the CVM and a gateway
to send traffic to.

We can also spin up a local DHCP server e.g. dnsmasq ...

Bug: b/344012112
Change-Id: If0f9551b74ab10673d2f53cac14ed62087311468