Use KeyFetcherWrapper in TVS for all platform type (local and GCP).
Instead of the untrusted part fetching the secret, it passes the authenticated users if any (right now I'm passing *default*) but we should pass the user from IAM or so ..

The rust wrapper acts as a factory that creates one fetcher and returns
the same instance all the time. This is the best thing we can do when mixing
C++ and rust. Alternatively, we can have the untrusted part pass in a pointer
or a reference but we want to reduce passing unsafe objects as much as possible.

We also provide a function to register an echo KeyFetcher for unit-tests.
The function allows registration only once and if there is no KeyFetcher registered
already.

Bug: b/344012112
Change-Id: Icc18b302f506c7ad79d2d1c4e1970017e626dd2f