Enable storing and fetching appraisal policies in Spanner for GCP deployments.
This CL introduces a local and GCP policy fetchers similar to the key fetchers.
Also extend database_main CLI to store a policy from a file to Spanner.
Policies are stored in Spanner as serialized bytes as we are treating them as blobs.
This should lay the ground of supporting multiple policies.

Bug: b/344012112
Change-Id: I2d2f1187ca278428fb28f5372e5695e61b717223