Orchestrator to export GetHpkeKey() to the workload to pass in the key obtained from the TVS.
This CL, also provides a client implementation to be used by the workload.
I temporarily removed other services exported via UDS as we don't have visibility to the packages (and we are not using them). I will put them back soon.

I also noticed that Oak submodule were updated which broke our scripts that builds the system image.
I will fix it in the next CL so other folks can test in CVMs.
Bug: b/351007909
Bug: b/351010994

Change-Id: I048b67af067a371323384d623a75c96229a7f0f5