Create a system image for hats based on Oak's one.
We need our own image so that we use our orchestrator. This a replacement
to some shell scripts we have.

Note, that the oak_containters_syslogd bazel rules are disabled due to linkage
issues. I got it to work on my workstation but it's not portable as it
needs certain libraries installed. For that reason I added the folder under
.bazelignore. I will remove it from the ignore file once I have a better fix.

Bug: b/351007909
Change-Id: I3bf51aeddff1660d2f1ef1a71dc4345a64ad3f2b