Patch Oak to use the correct network interface name in the passed kernel cmdline.
I have already had the correct interface passed in our qemu.cc; but doing it in Oak while we are transitioning to ours.
In Oak's system image things work fine since the network interface is configured to come up, while in our case
we don't want it to come up as we have different networking modes.

Bug: b/351007909
Bug: b/351010994
Change-Id: I3271fa5dc977a979044261233960b0a1881a70cb