Add a patch to make the system image work in linux machines and remove it from .bazelignore.
oak_containers_systemd links against libsystemd.so however, in most systems the library
has a suffix to it e.g. libsystemd.so.0.88 or so. The compiler will only work if you install libsystemd-dev
or rename the file or symlink it to libsystemd.

So in this CL we are creating a build script that symlink the file in the OUT_DIR so that rust compiler can find it.

I excluded oak_containers_systemd from the hats_system_image build rule unless enable_syslogd is set to true:
```
bazel build //client/system_image:hats_system_image --//:enable_syslogd=true
```

I did this even though the patch works in linux system but just in case.

The correct solution I believe is to provide a build environment that has the library e.g. nix script or so.

Bug: b/351007909
Bug: b/351010994
Change-Id: Id2d64f42786dbd080c609db155385af21192a9c8