Use aspect hermetic C/C++ toolchain in Hats repo for reproducibility & portability.
This is to enable running binaries compiled in our engineers workstations in our testing machines.
For more information see: https://blog.aspect.build/hermetic-c-toolchain
Oak is also using the same toolchain.

The toolchain deploys an older version of libc/libc++ and other libraries.
This is so that binaries can work on machines with newer libraries (libraries ensure backward compatilibity).
Note, this is just used for compilation, we are not linking against the libraries statically.

One note about oak_containers_syslogd: the binary links against libsystemd.so which doesn't exist in the aspect sysroot.
We deploy oak_containers_syslogd with our system image, and currently we don't have a good way of compiling the binary.

Currently, we don't include it in our system image, unless the user explicitly say so by enabling a flag.
For building our system image we rely on a shellscript that uses cargo and put it in a place where bazel can find the binary (disabled by default).

Alternative non-feasible options:
1. Precompile libsystemd.so (and other dependencies e.g. libpcap, libgcrypt .. etc). However, this would require checking in
a number of binaries (i.e ELF) files.
2. Depends on all the libraries and compile them everytime.
3. Use system installed libraries (we have a patch that makes this work, users can disable the hermetic toolchain and compile the binary).

Bug: b/361596941
Bug: b/351007909
Change-Id: I8c406ba8440f7831e45f9a831d089e13136c7629