Update RegisterUser operation to generate a new secret for a user if they are already registered.
Note: Currently a user can only have 2 secrets stored at a time.

I was using this operation locally to test my implementation of the test client that it could recover multiple user secrets from multiple tvs's, but this might be helpful for Kritika to test her locking logic when generating secrets and encrypting them in gcp for a user.

Change-Id: I4b62dc88283444ceb8e0267c814ad4c02730ca5e