Upgrade oak dependency to the latest.
I plan to send few changes to Oak in order to enable some services exported by the orchestrator.

I also update the oak submodule to point at the same commit used by WORKSPACE to build stage0, and stage1. As the need to generate the evenlogs that is introduced in the new orchestrator.
Note stage1 is not bazelfied yet, and in theory we can use bazel to build stage0; however, we still need
do some work to setup the right toolchain settings.

Bug: b/344012112
Bug: b/351007909
Change-Id: Ib33ce69220e256a6fb742d39245a404f9d4587f6