Update secret sharing code to tell orchestrator if secret is doubly encrypted.
I am also updating the database in this cl to take a dek so that I can (in the next cl) test that I can decrypt the second layer in the test client.

Note: I am not storing the user_dek because this isn't going to be in our database, when we initialize the launcher we provide it the private key encryption key so that it can then be fetched with FetchOrchestratorMetadata so that we can do the second layer decryption in the orchestrator.

Bug: b/365178747

Change-Id: I121b8fbc3081be7a76ebe88bf8a221eb184c14da