Fix move after use in launcher main.
LauncherConfig proto is already moved when addr_uri is constructed.
This makes the addr_uri and vsock_uri to be invalid (the behavior is undefined).
I will restructure the config in a later CL and do extra validation, but for now I'm generating the URIs
in the HatsLauncher::Create() and have HatsLauncher store them as members instead
of asking the client to pass them when calling Start().

Bug: b/351007909
Change-Id: I9601893714f2d03049f7544d092df7d14af6e8ae