Assume that the key provisioning role is leader instead of hard coding it in the launcher.
With the new launcher: the orchestrator panics when it tries to talk to the key provisining service because there isn't any. Instead of implementing one that returns UNIMPLEMENTED I changed the logic where the role is *leader*.

There are two roles: leader and follower.
Oak hasn't implemented follower roles and their launcher service returns *Leader* always [1].
In our case instead of implementing a service that *Leader* always and implement GetGroupKey
to return unimplemented, we remove the follower logic from the code.

[1] https://github.com/project-oak/oak/blob/385d5d40f9da5f0ab1df8d4e3ddfa0062b813490/oak_containers/launcher/src/server.rs#L210
Bug: b/351007909

Change-Id: I3744c45c39284286fbbe333c6be3259ab555c7de