Remove the logic that fetches application configs from the orchestrator
We don't provide a way to load application configs currently, and the rpc is not implemented.
This makes the orchestrator fails.
also, make SendEvidence rpc handler returns Ok instead of unimplemented, causing the orchestrator to fail
with the new launcher.
This is so the user of code can implement the method if they want to access the evidence themselves.

Bug: b/351007909
Change-Id: I6f6e9e712ea8b6cad9571752e024b3537bd4ce08