Use file_descriptor from the outer scope if the log file name is not enabled.
If filename is not enabled, we are supposed to create a random file; however, the file descriptor
is not captured and we will end up passing -1 to posix_spawn_file_actions_adddup2.
Here I'm also printing the actual error message from posix_spawn_file_actions_adddup2 to help debugging
in the future.

Bug: b/351007909
Change-Id: Iff8919a16890406f78514cb39befc4bd33f8f1fb