Remove host_orchestrator_proxy_port from the config.
We don't have any service (in use) exported by the orchestrator to the launcher.
The only service exported is the key provisioning, which is not used currently.
If we decide to use it, we will use vsock instead (we have the CID of the guest).

This simplifies our configuration and reduces the number of port we have to handle
to avoid collisions (i.e. when multiple guests are running in the same host).

Bug: b/351007909
Change-Id: Ifbd8c7b03fa0e2bdd44ecdbb8eb0b35ad21557df