Have the OS choose vsock and tcp ports automatically, and only create TCP if the work loads needs it.
This simplifies the launcher config, as the only user to the exported services are the CVM.
For debugging `ss -a` or `lsof -Pi` can be used to find the ports.

Only create TCP if the workload needs Parc, and only create qemu forwarding rule if parc is enabled.
Have Grpc choose ports for us, it works out of the box with TCP but for vSock it's not well supported.

Grpc returns always 1 for vSock, and there is no way for trivial way to get the port or sockfd -- it always returns *1* [1].
We leverage SockMutator to get the socket file descriptor and use getsocketname to get the port.
Seems SO_PORTREUSE doesn't work well with vSock, otherwise we would bind the socket ourselves to reserve the port.

We can choose the port randomly, but there are chances that the port is already used.

[1] https://github.com/grpc/grpc/blob/f55bf225da0eafeeebffd507dcb57c625933d105/src/core/lib/address_utils/sockaddr_utils.cc#L371

Bug: b/351007909
Change-Id: I3c3540b95b336b13f6bda69512ff598dd462b6b6