Support patching python to register toolchain, for Kokoro
Docker builds run as root, which can't run python (used by pkg_tar)
Use builders submodule to register toolchain, which fixes it
However, this also hard-codes the python path, so the patch changes
this as needed.
The default is to not doing it at all for standard execution (bazel),
since the default path is for bazel-debian and set in the submodule.

Remove nopresubmit for pkg_tar builds, as they now work in Kokoro

Bug: b/351201455
Change-Id: I7ee4b938ff1f7c403e19b7ff0b1e97ce170c38e5