Add new appraisal policy representation as addition to the current proto.
I will remove the old fields, but it requires significant changes to the code.
So it will be in the next CL.
I updated kv.prototext as an example for the reviewer.

The new representation is specific to our use cases and is human readable and editable.
If one wants other representation, they can extend PolicyManager to accommodate that.

In the new policy, we don't allow skipping fields and we manually map it to Oak's referencevalues proto.

Also it would allow us to take the fields and store them in Spanner or other storage in our own representation. We will only use the proto to represent policies in files.

Bug: b/344012112
Change-Id: Ib2bb6eb39dbb74821a1cf6a5658fcaf316488964