Use a simplified appraisal policy version added in the previous CL.
The policy is both readable and editable by human and we only include fields we use.
We don't allow skipping fields arbitrary.
We still map it to Oak's ReferenceValues to use their verification functions.

Policy signature logic is modified to sign over the actual binary data instead of serialized protos.
Serialized protos are not consistent as they are dependent on the software that calcuates them (differs by version).
This would allow us to use any format to represent our policies.

Bug: b/344012112
Change-Id: I77561e3dcf376c4617f2c7cdc5412b89986f3bd9