Rename TrustedTvs to RequestHandler and put it in a crate called trusted_tvs.
TrustedTvs service is created everytime we are handling a request, so it makes sense to rename.
The more important reason, is that we creating a new module *Service* that is created during boot and spawn
a request handler for every request. The service object would hold appraisal policies and key materials.
This is to avoid proto de-serialization, and key parsing for every request.
We would do it once during boot time.

Also, we are planning to re-use the handler in other services running in a TEE.

Note that I used "_" in file names as rust_library doesn't know to translate "-" to "_" for modules in a crate.

Bug: b/344012112
Change-Id: I76c4f074715a926b6ac70d1a9250282029286f66