Leading zeros: prepend a label `hats` to the secret before splitting it.
Adding a label is one of the techniques used in Shamir secret [1].

The reason, I'm adding this label is that the leading zeros from the secret
will be gone after recovering. Either we change our client to add the right number of zeros.
Which is dependent on the encryption scheme and other variants.

In our current setup, the orchestrator will pass in the key without the leading zeros and the workload will fail.

Also, the tests are a little bit flaky, try running:
``$bazel test //crypto:secret-sharing_test --runs_per_test=1000```
You will see that some of the tests fail. Those are with keys with leading zeros.

Alternative solutions:
1) Prepend the original secret size; however, the array length is 64 bit and in few cases the secret will be larger than our constant prime.
2) Use proto or json to serialize the secret.

While, I'm touching this code, I did few cosmetic changes:
1) I removed the mut from the method signature (as they are not mutating any of the members).
2) I changed the public facing method signatures to take slices.
Passing by value forces the call site to clone the variable if the need to use it after the call.
Vector by reference forces the call site to use Vector, and they might use arrays or other containers.

[1] https://dark-crystal-java.gitlab.io/protocol-specification/shamirs-secret-sharing.html

Bug: b/344012112
Change-Id: I083d4f8a8ee40d614f564e4fcef3db878e2c01c5