Pass KeyFetcher to trusted tvs instead of calling functions that create a static fetcher.
This would enable more thorough testing and also allow us to replace the C++ key fetcher
with another one.
RequestHandler now takes a reference to a key provider trait, this would enable passing another
implementation of key fetcher (other than the wrapped C++ version).
This would be useful if we want to pass in an object that retrieve keys over grpc/microrpc (e.g. when run the trusted code in a low tcb tee).

Bug: b/344012112
Change-Id: I3e3dbec5ee49768e81c4532a0b5596fff3e2443d