Make policy-{manager|signature} work with no std.
This is so that we can reuse the components in a low tcb tee.
I switched to anyhow::Error as it works well in no std.
Also anyhow is nice if we want to propagate the full error messages by setting contexts (we are not doing so).
alternatively, we can use enums and (anyerror to generate strings for them or implement fmt trait for them).
I used vec and string from alloc (which are the same one reexported by std).

Bug: b/344012112
Change-Id: I67644498f3a2cca7f15cf4d74b160dc808f36fea