Add flag to log qemu to std
I removed this in the last cl when I was making this a test (we still have to run the script to build the oak containers + instability in the build are reasons we still don't have this as an automated test on kokoro).

We need the logs to be able to scrape the hashes to fix the appraisal policy. Alternatively you could just check the tmp log file as that it created and its location is printed out everytime we run ./test_main.

go/hats-simple-integration-testing

Bug: b/371300767
Change-Id: I7bba4399835ac04337b908659f0671c5c311add0