Update trusted application test to exit after 6 cycles of waiting.
On 4 machines I've tested with/without snp enabled the application takes no more than 30 seconds to start up.
Note: This is a temporary solution to help debug the tests locally, we have to make changes in the launcher to exit on such a failure because currently the launcher will just keep running regardless of if qemu launches successfully.

Bug: b/371300767
Change-Id: I09ac6fa63df6bb67b63b0ca205fdec9854089934