Add method in launcher to check on status of Qemu
We did not have a way to check on the status of qemu so in test_main we were seeing the test keep running even after qemu failed since the launcher didn't see the failure.

This gives us more useful logs while debugging as well as allows us to check on the status of qemu.

In the next cl I will make this a test case and add a readme to run the script before testing with bazel since the test expects binaries in certain places.

Bug: b/371300767
Change-Id: I4695d80c86c08253e3c54f76cb7ee2eaedf81b3a