Update Oak's dependencies to the latest.
The update helps in:
1) Resolve the issue with syslogd hash instability as Oak
fixed systemd.so in the toolchain sysroot.
This would help with the integration test.
I used bazel to build the syslogd instead of cargo so that
we use the *good* systemd.so from the sysroot.
In the near future, we should also update our toolchain so that
we use the system_image bazel rule without the need of pre-building
the syslogd.
2) Necessary visibility to build TVS for restricted kernel.

Bug: b/344012112
Change-Id: I592f8d390201ebc3c73c5af02ffcbdabad67f998