Fix system_bundle and runtime_bundle locations.
The test script now places the tar bundles in the test_data location
where we have bazel export these files to its runtime location and find
them there.

The system bundle will change as we make changes to the orchestrator but
we can have kokoro run the updated test script that places the system
bundle in test_data dir where bazel looks during runtime.

Note, we still do not have an automated way of updating the system
bundle and runtime bundle hashes in the appraisal policy so for now we
will do this manually.

Bug: b/371300767

Change-Id: I1115cccddc1e583a3bae3cc4409f1d47346e8be3