Remove TVS key conversion logic from key-provider trait.
We are putting back the conversion logic to the trusted tvs code.

This way we can avoid depending on the enclave/crypto library
in the key-provider.
We want to depend on a non-std crypto implementation for TVS in restricted kernel.
We cannot do that if the key-provider depends on another one.
Alternative solution is to use bazel flag to change the dependency or make P256Scalar as a trait
but then we would need our code as traits do not have known size during compile time so we have to
use boxes or other heap allocated variables.

Bug: b/344012112
Change-Id: Id5ade5520dd54a5584fcc9eef81468fecf591807