Make trusted TVS takes KeyProvider trait instead of KeyFetcher.
This way we can pass it arbitrary objects that implements the trait.
RequestHandler already takes a reference of the trait.
This would allow us to reuse both service.rs and request_handler.rs for
TVS enclave (in Oak's restricted kernel).

Note that we have to use dynamically allocated objects (Box) as the object
size s not known during compile time.

We will provide a different implementation of KeyProvider for TVS enclave.

Bug: b/344012112
Change-Id: I13ec0d82d5660775c32552ad9ff5f2cb61469752