Use Arc instead of references for objects passed to request_handler.
Earlier, Service object owned keyfetcher, policy manager and others.
Service creates a request handler and passes the aforementioned objects
by reference.
Rust does not support self referential structs. This mean Service cannot
own RequestHandler object. In C++, we pass RequestHandlers to C++ code
and we make sure that Service object out-lives all RequestHandlers.

For TVS in restricted kernel we would like to have Service to own RequestHandler
and this is not possible as self referential structs are not supported.
So we use Arc instead of references.

Side benefit of using Arc, is that we don't have to worry
that the C++ code make sure that Service outlives RequstHandlers.

Bug: b/344012112
Change-Id: I537c6e5d39f5ad1c7cd736116d7d0bde43d9e48b