Add integration test for TVS in split mode (untrusted and trusted).
The integration test is a binary to avoid running it when calling bazel test //...
We could make it a test target and tag it with virtualization; however, this is an integration test
and we don't expect it to run with all other tests.

The test requires a number of binaries (TVS enclave app, stage0, restricted kernel, and initrd).
In order to run the test, one need to run the pre-build the aforementioned binaries.
The binaries need to be in a location that exports the files so I created binaries/ subfolder
added BUILD and README.md to it. I modified .gitignore to ignore all other files.

Bug: b/344012112
Change-Id: I88308261865c362138d57564838b78026047e61f