Add GetLatestNPoliciesForDigest() to AppraisalPolicy fetcher.
The method allows retrieving policies for a given application digest.
The method allow us to:
1. Retrieve policies for a particular hash instead of getting N latest policies.
2. Writing a policy fetcher that gets policies when TVS receives a request
   instead of fetching N policies during boot time.
3. This allows to retrieve the latest polices and not to wait for task reboots.

This methods becomes handy if we have a large number of policies as we expect
each application digest to have multiple different policies as the ACPI and memory map
measurements change with various VM configurations (memory and number of cores).

Hopefully, we don't need to add a bottlenick to spanner for every request.
yes, we query spanner for user secrets and authentication keys, however,
policies are much larger than user data.

Bug: b/344012112
Change-Id: Ic0501df6a842ea46b960caef253a3335b80d4926