Add an option to enable dynamic policy fetching mode in standalone TVS.
In this mode, policies are fetched dynamically for every request.
When an attestion report is received, TVS queries PolicyFetcher (local or Spanner0
for policies matching the application layer digest, and check measurements against those.

This feature is guarded behind a flag, and it might perform a bit worse that the default mode
(fetching policies and pre-process them during boot time).
However, quick experiment didn't show that the performance is bad.
Although for every request, we fetch the storage for policies, de-serialize the policies,
and check the signatures against them.

We might need this feature if:
1. We have a large number of policies (if we needed to support large number of work load).
2. We need policies to be updated instantly (and not to wait for the next task reboot).

This feature could take a lot of optimization, e.g. caching the policies.

Bug: b/344012112
Change-Id: I11f6c671e978670721dac4199b57811634be746d