Move the logic to download the tee certificate (vcek) early in the launching process.
Tee certificate is downloaded when the orchestrator asks the launcher for it.
Most often AMD website times out on us and it's not clear what went wrong.
One need to look at the CVM, in particular orchestrator log to know what went wrong.
In fact, there is no point in launching a CVM if we cannot get the certificate.

In this change, I moved the logic of downloading the certificate from launcher-server
to launcher class. Launcher passes the certificate to launcher-server.

This way we can add retries mechanisms before launching the CVM.

Bug: b/351007909
Change-Id: I13b3a6a29a4336b49991f570e305507fff28a586