Select one launcher to download the tee certificate instead of each one downloading a clone.
Store tee certificate in an agreed locations so that we only download it once.

KDS rates limits us if we our traffic increases a certain threshold.
We observed this when we started multiple launcher at the same time.
Further, KDS does not offer any SLO or guarantees it's on best effort.

This CL introduces and ReadOrDownloadCertificate() to replace DownloadCertificate().

The function tries to read tee certificate from a known file location, if no file exists
it downloads the certificate from the vendor's key distribution portal.
The function uses advisory locking on the agreed certificate file for
synchronization across launchers in the machine. The first launcher that
grabs the lock when the is empty, downloads the certificate while the rest
wait for it. This way, we download the certificate only once.

Change-Id: I06e8d66bea749236135e337b1c899a9bbe965292