Append the current user to the certificate file name.
This way we don't need to make the file world readable/writable.

If two users are using the same machiens to launch CVMs then one
might not have access to open the certificate file.
Here, we create a certificate file for each use rto avoid giving
wide access to the file. In practise giving wide access should be ok
as the certificate is public, but usually world writable files triggers
suspicions.

Bug: b/351007909
Change-Id: I4e6d592bd2c06b1088857209e87a8adc4bd14a1c