Check if launcher->IsAppReady() after we exit the while loop{}.
Exiting the loop and checking the launcher status does not guarantee
that the app is ready.
Multiple times I faced an issue where the test fails due to the app
taking longer than expected. In this case, we shouldn't continue and
print an obsecure error message.

I will change IsAppReady() to take a call back instead  in an upcoming CL.

Bug: b/408055875
Change-Id: I190b084fbe3ba5256dde8327d4a1691977b28c2e