Add another appraisal policy for ACPI different order.
Apparently, newer version of QEMU started adding another entry to
the XDST table and if the order is flipped the hash changes.
I will propose a different way to measure the digest in stage0 to
Oak and I will see if we can tweak qemu parameters with specific
ACPI table settings.

I added a policy for insecure and for one of the configuration
(the one we are using by default). For now, I will update soon.

Bug: b/404822854
Change-Id: If7013fcf8a60f111769a4eb7587041085eb96bff