Create BUILD file to export oak_container_syslogd and disable checking the artificats.
Our system_image links against the oak_contaier_syslogd in the prebuilt directory
and that requires a BUILD file to export the file.
We create the BUILD file when we call build_oak_containers_syslogd(); but we don't
when we copy the artifiact.

Also, I disabled checking the artifacts as kokoro expect oak submodule to be in a different
directory.

Bug: b/408055875
Change-Id: I4b14e3729f2583d506c6695aaf811a164b178d6c