Fix flaky otel-counter test
There are two objects added to the collected_metrics vector
The first comes from the library/system, the second is the OtelCounter

The first one is flakey, and doesn't always attach in time for the
assert that the length is 2. The existence/behavior of the this one is
not in the scope of this test.

Instead, take the last item in the collected metrics, OtelCounter

Change-Id: I4d3351a4b113c62b077f9750917ea90e47a081db