Appraisal policy CLI: Calculate memory map digest.
Stage0 gets E820 table from QEMU and then modifies it.
The code in this CL follows the same path, it builds
the table the way QEMU does, and then modifies it the
way stage0 does.

The structs that represents the table entries and the
logic to add and merge memory regios were cloned from
Oak's code with minor modification.
We should use their code, however, we don't have visibility.
I will send them a CL to enable us to link directly against theirs.

Bug: b/404822854
Change-Id: I5e7d2b3c9699017bce3888226d362a17ecc33802