Clone https://github.com/LIT-Protocol/sev-snp-utils.git to third_party.
We want to use the library to calculate stage0 measurements.
The library as is doesn't work for us for the following reasons:
a) stage0 sections are not pointer aligned: the library uses bytemuck
to convert array of bytes to structs and the library requires
pointers to be aligned (i.e. % 4 == 0) and in Oak's binary
the location of the omvf and snp headers are not in aligned location.

b) Openssl dependencies: the library uses openssl which doesn't compile with our aspect tool chain as it requires shared libraries from openssl. In theory we can compile Openssl in bazel for the crate similar to [1]; however, we are already using boringssl and this would increase the compilation time.

I'm cloning the library versus copying the code as it would be easier to give credit to the original code. We could use patches; however, the patch will be hard to read.

We could use snp_measurement but that requires some work for the following:
a) Stage0 measurements does not match the one we get from our CVM (either
there is an issue or that it's producing measurements for different vmm/hypervisor).

b) The crate is a binary and not a library.

[1] https://github.com/ouillie/bazel-rust-musl/blob/main/MODULE.bazel

Bug: b/404822854
Change-Id: I116b3965b6ee6fd0475f3f263a7ff23676f2e90c