Update Oak dependency to d83ae4deadebd2a058af3bb8b91cad4d2925f81c.
Multiple dependencies needed to be updated in order to work with
new Oak versions. I had to update rust and move crate specs to modbzl
and do some minor changes to the code base.

This CL is large since rust and oak dependencies needed to be updated
together (along with other dependencies).

TODO in upcoming CLs:
1) Update digests in Genoa policies.
2) Fix the using deprecated things warnings (due to upgrade).
3) Fix build in the SNP presubmit server.

Bug: b/435176862
Change-Id: I93f02a61c8f0e31db8dab0540a86527764c78fb0