Introduce BinaryStore trait for OVMF binary lookup
Notes:
The dynamic_attestation feature requires TVS to store and be able to
lookup the OVMF binary files from the dynamic appraisal policy's
stage0_ovmf_binary_hash field.
 - The trait only needs one method get_ovmf_binary which uses sha384 as
   a key to look for a raw byte slice
 - Includes a test-only implementation for the trait which uses a hashmap for testing ease
 - Unit test to verify the lookup is working.
 - Moves dynamic_attestation flag to root for new
   BinaryStore trait, updating BUILD files accordingly.
This change is to prep for dynamic computation of the stage0 measurement "expected_hash", in
which we start with the OVMF binary as our "partial hash".

Bug: b/434016988
Change-Id: I1d35028f4f313d3dded84a36387828daf52ae09c